Module xi_core_lib::config

Structs

BufferItems

The concrete type for buffer-related settings.

Config

A frozen collection of settings, and their sources.

ConfigManager

ConfigPair

Represents the common pattern of default settings masked by user settings.

Enums

ConfigDomain

A ConfigDomain describes a level or category of user settings.

ConfigDomainExternal

The external RPC sends ViewIds, which we convert to BufferIds internally.

ConfigError

The errors that can occur when managing configs.

Type Aliases

BufferConfig

Table

A map of config keys to settings