Struct xi_core_lib::core::WeakXiCore

pub struct WeakXiCore(_);

A weak reference to the main state. This is passed to plugin threads.

Implementations

impl WeakXiCore

pub fn plugin_connect(&self, plugin: Result<Plugin, Error>)

Called immediately after attempting to start a plugin, from the plugin’s thread.

pub fn plugin_exit(&self, plugin: PluginId, error: Result<(), ReadError>)

Called from a plugin runloop thread when the runloop exits.

pub fn handle_plugin_update(
&self,
plugin: PluginId,
view: ViewId,
response: Result<Value, RpcError>
)

Handles the result of an update sent to a plugin.

All plugins must acknowledge when they are sent a new update, so that core can track which revisions are still ‘live’, that is can still be the base revision for a delta. Once a plugin has acknowledged a new revision, it can no longer send deltas against any older revision.