Struct xi_core_lib::find::Find

pub struct Find { /* fields omitted */ }

Contains logic to search text

Implementations

impl Find

pub fn new(id: usize) -> Find

pub fn id(&self) -> usize

pub fn occurrences(&self) -> &Selection

pub fn hls_dirty(&self) -> bool

pub fn find_status(
&self,
view: &View,
text: &Rope,
matches_only: bool
) -> FindStatus

pub fn set_hls_dirty(&mut self, is_dirty: bool)

pub fn update_highlights(&mut self, text: &Rope, delta: &RopeDelta)

pub fn unset(&mut self)

Unsets the search and removes all highlights from the view.

pub fn update_find(
&mut self,
text: &Rope,
start: usize,
end: usize,
include_slop: bool
)

Execute the search on the provided text in the range provided by start and end.

pub fn next_occurrence(
&self,
text: &Rope,
reverse: bool,
wrapped: bool,
sel: &Selection
) -> Option<SelRegion>

Return the occurrence closest to the provided selection sel. If searched is reversed then the occurrence closest to the start of the selection is returned. wrapped indicates that if the end of the text is reached the search continues from the start.