Struct xi_core_lib::layers::Layers

pub struct Layers { /* private fields */ }

A collection of layers containing scope information.

Implementations

impl Layers

pub fn get_merged(&self) -> &Spans<Style>

pub fn add_scopes( &mut self, layer: PluginPid, scopes: Vec<Vec<String>>, style_map: &ThemeStyleMap )

Adds the provided scopes to the layer’s lookup table.

pub fn update_all(&mut self, delta: &RopeDelta)

Applies the delta to all layers, inserting empty intervals for any regions inserted in the delta.

This is useful for clearing spans, and for updating spans as edits occur.

pub fn update_layer( &mut self, layer: PluginPid, iv: Interval, spans: Spans<u32> )

Updates the scope spans for a given layer.

pub fn remove_layer(&mut self, layer: PluginPid) -> Option<ScopeLayer>

Removes a given layer. This will remove all styles derived from that layer’s scopes.

pub fn theme_changed(&mut self, style_map: &ThemeStyleMap)

pub fn debug_print_spans(&self, iv: Interval)

Prints scopes and style information for the given Interval.