Module xi_core_lib::line_cache_shadow

Data structures for tracking the state of the front-end’s line cache and preparing render plans to update it.

Structs

Builder

Builder for LineCacheShadow object.

LineCacheShadow

The line cache shadow tracks the state of the line cache in the front-end. Any content marked as valid here is up-to-date in the current state of the view. Also, if dirty is false, then the entire line cache is valid.

PlanIterator

PlanSegment

RenderPlan

Span

Enums

RenderTactic

Constants

ALL_VALID

CURSOR_VALID

INVALID

STYLES_VALID

TEXT_VALID