Module xi_core_lib::rpc

The main RPC protocol, for communication between xi-core and the client.

We rely on Serde for serialization and deserialization between the JSON-RPC protocol and the types here.

Structs

EditCommand

A helper type, which extracts the view_id field from edit requests and notifications.

FindQuery

LineRange

An inclusive range.

MouseAction

A mouse event. See the note for LineRange.

Position

Enums

CoreNotification

The notifications which make up the base of the protocol.

CoreRequest

The requests which make up the base of the protocol.

EditNotification

The edit-related notifications.

EditRequest

The edit related requests.

GestureType

An enum representing touch and mouse gestures applied to the text.

PluginNotification

The plugin related notifications.

SelectionGranularity

The smallest unit of text that a gesture can select

SelectionModifier

Represents how the current selection is modified (used by find operations).