Enum xi_core_lib::rpc::EditNotification

pub enum EditNotification {

    Insert {
        chars: String,
    },
    Paste {
        chars: String,
    },
    DeleteForward,
    DeleteBackward,
    DeleteWordForward,
    DeleteWordBackward,
    DeleteToEndOfParagraph,
    DeleteToBeginningOfLine,
    InsertNewline,
    InsertTab,
    MoveUp,
    MoveUpAndModifySelection,
    MoveDown,
    MoveDownAndModifySelection,
    MoveLeft,
    MoveBackward,
    MoveLeftAndModifySelection,
    MoveRight,
    MoveForward,
    MoveRightAndModifySelection,
    MoveWordLeft,
    MoveWordLeftAndModifySelection,
    MoveWordRight,
    MoveWordRightAndModifySelection,
    MoveToBeginningOfParagraph,
    MoveToBeginningOfParagraphAndModifySelection,
    MoveToEndOfParagraph,
    MoveToEndOfParagraphAndModifySelection,
    MoveToLeftEndOfLine,
    MoveToLeftEndOfLineAndModifySelection,
    MoveToRightEndOfLine,
    MoveToRightEndOfLineAndModifySelection,
    MoveToBeginningOfDocument,
    MoveToBeginningOfDocumentAndModifySelection,
    MoveToEndOfDocument,
    MoveToEndOfDocumentAndModifySelection,
    ScrollPageUp,
    PageUpAndModifySelection,
    ScrollPageDown,
    PageDownAndModifySelection,
    SelectAll,
    AddSelectionAbove,
    AddSelectionBelow,
    Scroll(LineRange),
    Resize(Size),
    GotoLine {
        line: u64,
    },
    RequestLines(LineRange),
    Yank,
    Transpose,
    Click(MouseAction),
    Drag(MouseAction),
    Gesture {
        line: u64,
        col: u64,
        ty: GestureType,
    },
    Undo,
    Redo,
    Find {
        chars: String,
        case_sensitive: bool,
        regex: bool,
        whole_words: bool,
    },
    MultiFind {
        queries: Vec<FindQuery>,
    },
    FindNext {
        wrap_around: bool,
        allow_same: bool,
        modify_selection: SelectionModifier,
    },
    FindPrevious {
        wrap_around: bool,
        allow_same: bool,
        modify_selection: SelectionModifier,
    },
    FindAll,
    DebugRewrap,
    DebugWrapWidth,
    DebugPrintSpans,
    DebugToggleComment,
    Uppercase,
    Lowercase,
    Capitalize,
    Reindent,
    Indent,
    Outdent,
    HighlightFind {
        visible: bool,
    },
    SelectionForFind {
        case_sensitive: bool,
    },
    Replace {
        chars: String,
        preserve_case: bool,
    },
    ReplaceNext,
    ReplaceAll,
    SelectionForReplace,
    RequestHover {
        request_id: usize,
        position: Option<Position>,
    },
    SelectionIntoLines,
    DuplicateLine,
    IncreaseNumber,
    DecreaseNumber,
    ToggleRecording {
        recording_name: Option<String>,
    },
    PlayRecording {
        recording_name: String,
    },
    ClearRecording {
        recording_name: String,
    },
    CollapseSelections,

}

The edit-related notifications.

Alongside the EditRequest members, these commands constitute the API for interacting with a particular window and document.

Variants

Insert

Fields of Insert

chars: String

Paste

Fields of Paste

chars: String

DeleteForward

DeleteBackward

DeleteWordForward

DeleteWordBackward

DeleteToEndOfParagraph

DeleteToBeginningOfLine

InsertNewline

InsertTab

MoveUp

MoveUpAndModifySelection

MoveDown

MoveDownAndModifySelection

MoveLeft

MoveBackward

MoveLeftAndModifySelection

MoveRight

MoveForward

MoveRightAndModifySelection

MoveWordLeft

MoveWordLeftAndModifySelection

MoveWordRight

MoveWordRightAndModifySelection

MoveToBeginningOfParagraph

MoveToBeginningOfParagraphAndModifySelection

MoveToEndOfParagraph

MoveToEndOfParagraphAndModifySelection

MoveToLeftEndOfLine

MoveToLeftEndOfLineAndModifySelection

MoveToRightEndOfLine

MoveToRightEndOfLineAndModifySelection

MoveToBeginningOfDocument

MoveToBeginningOfDocumentAndModifySelection

MoveToEndOfDocument

MoveToEndOfDocumentAndModifySelection

ScrollPageUp

PageUpAndModifySelection

ScrollPageDown

PageDownAndModifySelection

SelectAll

AddSelectionAbove

AddSelectionBelow

Scroll(LineRange)

Resize(Size)

GotoLine

Fields of GotoLine

line: u64

RequestLines(LineRange)

Yank

Transpose

Click(MouseAction)

Drag(MouseAction)

Gesture

Fields of Gesture

line: u64col: u64ty: GestureType

Undo

Redo

Find

Fields of Find

chars: Stringcase_sensitive: boolregex: boolwhole_words: bool

MultiFind

Fields of MultiFind

queries: Vec<FindQuery>

FindNext

Fields of FindNext

wrap_around: boolallow_same: boolmodify_selection: SelectionModifier

FindPrevious

Fields of FindPrevious

wrap_around: boolallow_same: boolmodify_selection: SelectionModifier

FindAll

DebugRewrap

DebugWrapWidth

DebugPrintSpans

Prints the style spans present in the active selection.

DebugToggleComment

Uppercase

Lowercase

Capitalize

Reindent

Indent

Outdent

HighlightFind

Indicates whether find highlights should be rendered

Fields of HighlightFind

visible: bool

SelectionForFind

Fields of SelectionForFind

case_sensitive: bool

Replace

Fields of Replace

chars: Stringpreserve_case: bool

ReplaceNext

ReplaceAll

SelectionForReplace

RequestHover

Fields of RequestHover

request_id: usizeposition: Option<Position>

SelectionIntoLines

DuplicateLine

IncreaseNumber

DecreaseNumber

ToggleRecording

Fields of ToggleRecording

recording_name: Option<String>

PlayRecording

Fields of PlayRecording

recording_name: String

ClearRecording

Fields of ClearRecording

recording_name: String

CollapseSelections