Struct xi_core_lib::selection::SelRegion

pub struct SelRegion {
    pub start: usize,
    pub end: usize,
    pub horiz: Option<HorizPos>,
    pub affinity: Affinity,
}

A type representing a single contiguous region of a selection. We use the term “caret” (sometimes also “cursor”, more loosely) to refer to a selection region with an empty interior. A “non-caret region” is one with a non-empty interior (i.e. start != end).

Fields
start: usize

The inactive edge of a selection, as a byte offset. When equal to end, the selection range acts as a caret.

end: usize

The active edge of a selection, as a byte offset.

horiz: Option<HorizPos>

A saved horizontal position (used primarily for line up/down movement).

affinity: Affinity

The affinity of the cursor.

Implementations

impl SelRegion

pub fn new(start: usize, end: usize) -> Self

Returns a new region.

pub fn caret(pos: usize) -> Self

Returns a new caret region (start == end).

pub fn with_horiz(self, horiz: Option<HorizPos>) -> Self

Returns a region with the given horizontal position.

pub fn with_affinity(self, affinity: Affinity) -> Self

Returns a region with the given affinity.

pub fn min(self) -> usize

Gets the earliest offset within the region, ie the minimum of both edges.

pub fn max(self) -> usize

Gets the latest offset within the region, ie the maximum of both edges.

pub fn is_caret(self) -> bool

Determines whether the region is a caret (ie has an empty interior).

pub fn is_upstream(self) -> bool

Determines whether the region’s affinity is upstream.