Struct xi_core_lib::selection::Selection

pub struct Selection { /* fields omitted */ }

A set of zero or more selection regions, representing a selection state.

Implementations

impl Selection

pub fn new() -> Selection

Creates a new empty selection.

pub fn new_simple(region: SelRegion) -> Selection

Creates a selection with a single region.

pub fn clear(&mut self)

Clear the selection.

pub fn collapse(&mut self)

Collapse all selections into a single caret.

pub fn search(&self, offset: usize) -> usize

pub fn add_region(&mut self, region: SelRegion)

Add a region to the selection. This method implements merging logic.

Two non-caret regions merge if their interiors intersect; merely touching at the edges does not cause a merge. A caret merges with a non-caret if it is in the interior or on either edge. Two carets merge if they are the same offset.

Performance note: should be O(1) if the new region strictly comes after all the others in the selection, otherwise O(n).

pub fn regions_in_range(&self, start: usize, end: usize) -> &[SelRegion]

Gets a slice of regions that intersect the given range. Regions that merely touch the range at the edges are also included, so it is the caller’s responsibility to further trim them, in particular to only display one caret in the upstream/downstream cases.

Performance note: O(log n).

pub fn delete_range(&mut self, start: usize, end: usize, delete_adjacent: bool)

Deletes all the regions that intersect or (if delete_adjacent = true) touch the given range.

pub fn add_range_distinct(&mut self, region: SelRegion) -> (usize, usize)

Add a region to the selection. This method does not merge regions and does not allow ambiguous regions (regions that overlap).

On ambiguous regions, the region with the lower start position wins. That is, in such a case, the new region is either not added at all, because there is an ambiguous region with a lower start position, or existing regions that intersect with the new region but do not start before the new region, are deleted.

pub fn apply_delta(
&self,
delta: &RopeDelta,
after: bool,
drift: InsertDrift
) -> Selection

Computes a new selection based on applying a delta to the old selection.

When new text is inserted at a caret, the new caret can be either before or after the inserted text, depending on the after parameter.

Whether or not the preceding selections are restored depends on the keep_selections value (only set to true on transpose).