Module xi_core_lib::styles

Management of styles.

Structs

Style

A mergeable style. All values except priority are optional.

ThemeSettings

Various properties meant to be used to style a text editor. Basically all the styles that aren’t directly applied to text like selection color. Use this to make your editor UI match the highlighted text.

ThemeStyleMap

A map from styles to client identifiers for a given Theme.

Constants

DEFAULT_THEME

N_RESERVED_STYLES