Module xi_core_lib::width_cache

Cache and utilities for doing width measurement.

Structs

CodepointMono

A measure in which each codepoint has width of 1.

WidthBatchReq

A batched request, so that a number of strings can be measured in a a single RPC.

WidthCache

WidthReq

A request for measuring the widths of strings all of the same style (a request from core to front-end).

Traits

WidthMeasure

A trait for types that provide width measurement. In the general case this will be provided by the frontend, but alternative implementations might be provided for faster measurement of ‘fixed-width’ fonts, or for testing.

Type Definitions

Token

A token which can be used to retrieve an actual width value when the batch request is submitted.

WidthResponse

The response for a batch of WidthReqs.