Trait xi_core_lib::width_cache::WidthMeasure

pub trait WidthMeasure {
    // Required method
    fn measure_width(
        &self,
        request: &[WidthReq]
    ) -> Result<WidthResponse, Error>;
}

A trait for types that provide width measurement. In the general case this will be provided by the frontend, but alternative implementations might be provided for faster measurement of ‘fixed-width’ fonts, or for testing.

Required Methods

fn measure_width(&self, request: &[WidthReq]) -> Result<WidthResponse, Error>